requirements.txt
discord.py
google-generativeai
edge-tts
moviepy
imageio-ffmpeg